Up to latest Ruby 1.8.7, latest Rubygems and Chef
This Rubygems upgrade means a ton of deprecation warnings, but it fixes
the "invalid date format in specification" bug that we were going to
keep tripping over as we upgraded Selenium and its dependencies.

Chef was only held at an older revision to keep a json gem at an old
enough release that it wouldn't trigger the same Rubygems bug, so may as
well go to the latest.